+++ src/App.js
@@ -7,7 +7,6 @@ import Routes from './routes'
 import  Header  from './components/NavBar'
 import Footer from './components/Footer';
 
-
 function App() {
 
   return (
