+++ src/components/Footer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../css/footer.css';
+
 
 const Footer = () => {
   return (
