+++ src/config/firebase.js
