+++ src/css/style.css
@@ -205,4 +205,45 @@ margin-left: 1em;
 
 }
 
+/* Footer */
+.footer {
+  padding: 2px 10px;
+  align-items: center;
+  background-color: rgb(30, 52, 69);
+  color: rgb(255, 255, 255);
+}
 
+.row {
+  display: flex;
+  justify-content: space-between;
+}
+
+.icon {
+  color: #fff;
+  margin-right: 20px;
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+}
+
+.icon div {
+  display: flex;
+}
+
+.icon a img {
+  width: 50px;
+  height: 50px;
+  padding: 2px;
+}
+
+.text-size {
+  font-size: 12px;
+  line-height: 0.3;
+}
+
+.copyright {
+  text-align: center;
+  padding-bottom: 5px;
+}
+
+/* end Footer */
